Consumer Code Support
This adds support to send consumer codes to the host device.  As part of this,
we are also changing how we determine modifier keys and simplifying the process
overall on detecting different key types.

We are also changing key codes from `uint16_t` to `uint8_t` variable type, as
there is no need really to have a 16bit integer value for this.

As for Consumer Codes, we use range A5-C8 to add support for these directly, and
then translate these using a Macro definition to the actual HID codes needed for
these to work.

TODO:

I still need to add support to toggle between pre-defined layers for the keyboard
as well as adding an emulated NumLock support method, as when connected to an
Apple device, NumLock is not natively supported on this platform.

+++ src/keymap.c
@@ -20,7 +20,7 @@
 
 #include "keymap.h"
 
-uint16_t keymap_get_key_val(int layer, uint8_t pos) {
+uint8_t keymap_get_key_val(int layer, uint8_t pos) {
   uint8_t row = (pos >> 4) & 0x0F;
   uint8_t col = pos & 0x0F;
   return keymap_map[layer][row][col];
